Add a button with logic to remove/create new grid
Decided to use a while loop within an event handler on the button as I
thought this would be the easiest way to control what the user was
entering (i.e. max of 100). Moved grid creation into its own function
to make it easier to call in the button event handler.

Future updates might be useful to throw everything in its own function,
but I don't think it is necessary for this project.