Add progressive darkening effect on div interactions
Decided to use a window property as I didn't know
a better way to grab the current opacity and add to
it with each interaction. I think this solves the
project, but go ahead and refactor if worth the
effort in the future.

Also, if you actually want to play this game, I
would recommend going back two commits before this
one because the extra credit details end up with a
poor user interface in my opinion.